Allow fractional logging frequencies.
This patch allows you to specify a float value for hz. The net
effect is that you can increase the logging interval to values
larger than 1.0 seconds, which is useful if you want to use
the profiling code to profile a larger timeframe.

I tested this with 0.1, 0.5, 22.5 values, and seems to work
as expected.

+++ bootchart.h
@@ -99,7 +99,7 @@ extern int initcall;
 extern int samples;
 extern int cpus;
 extern int len;
-extern int hz;
+extern double hz;
 extern int scale_x;
 extern int scale_y;
 extern int overrun;
